lambda_function_trigger/requirements.txt
boto3
requests
websocket-client
websockets
asyncio